(C) Remove some compilation warnings in the library.

+++ eclipse/plugins/net.sf.orcc.backends/runtime/C/orcc/src/accessFile.c
@@ -34,35 +34,35 @@
 #include <sys/stat.h>
 
 
-int openFile(char* filename) {
+long openFile(char* filename) {
 	FILE* fd = NULL;
 	fd = fopen(filename,"r+b");
 	if(fd == NULL) {
 		fprintf(stderr, "Error during opening of %s\n", filename);
 		exit(-2);
 	}
-	return (int)fd;
+	return (long)fd;
 }
 
-void writeByte(int desc, char value ) {
+void writeByte(long desc, char value ) {
 	FILE* fd = (FILE*) desc;
 
 	fwrite(&value, 1, 1, fd);
 }
 
-int closeFile(int desc) {
+int closeFile(long desc) {
 	FILE* fd = (FILE*) desc;
 	return fclose(fd);
 }
 
-int sizeOfFile(int desc){
+int sizeOfFile(long desc){
 	FILE* fd = (FILE*) desc;
 	struct stat st; 
 	fstat(fileno(fd), &st); 
 	return st.st_size;
 }
 
-unsigned char readByte(int desc){
+unsigned char readByte(long desc){
 	FILE* fd = (FILE*) desc;
 	unsigned char buf[1];
 	int n = fread(&buf, 1, 1,fd);
